+++ R/rm_white.R
@@ -52,7 +52,6 @@ rm_white <- hijack(rm_default, pattern = "@rm_white", clean = FALSE, trim = FALS
 #' or before a right bracket ("}", ")", "]").
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_bracket <- hijack(rm_default, pattern = "@rm_white_bracket")
@@ -63,7 +62,6 @@ rm_white_bracket <- hijack(rm_default, pattern = "@rm_white_bracket")
 #' combination of a colon, semicolon.
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_colon <- hijack(rm_default, pattern = "@rm_white_colon")
@@ -73,7 +71,6 @@ rm_white_colon <- hijack(rm_default, pattern = "@rm_white_colon")
 #' \code{rm_white_comma} - Remove white space before a comma.
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_comma <- hijack(rm_default, pattern = "@rm_white_comma")
@@ -83,7 +80,6 @@ rm_white_comma <- hijack(rm_default, pattern = "@rm_white_comma")
 #' \code{rm_white_endmark} - Remove white space before endmark(s) (".", "?", "!").
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 #' @author \code{rm_white_endmark}/\code{rm_white_punctuation} - \href{http://stackoverflow.com}{stackoverflow}'s hwnd and Tyler Rinker <tyler.rinker@@gmail.com>. 
@@ -96,7 +92,6 @@ rm_white_endmark <- hijack(rm_default, pattern = "@rm_white_endmark")
 #' \code{rm_white_lead} - Remove leading white space.
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_lead <- hijack(rm_default, pattern = "@rm_white_lead", clean = FALSE, 
@@ -107,7 +102,6 @@ rm_white_lead <- hijack(rm_default, pattern = "@rm_white_lead", clean = FALSE,
 #' \code{rm_white_lead_trail} - Remove leading or trailing white space.
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_lead_trail <- hijack(rm_default, pattern = "@rm_white_lead_trail", 
@@ -118,7 +112,6 @@ rm_white_lead_trail <- hijack(rm_default, pattern = "@rm_white_lead_trail",
 #' \code{rm_white_trail} - Remove trailing white space.
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_trail <- hijack(rm_default, pattern = "@rm_white_trail", 
@@ -130,7 +123,6 @@ rm_white_trail <- hijack(rm_default, pattern = "@rm_white_trail",
 #' white space).
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_multiple <- hijack(rm_default, pattern = "@rm_white_multiple")
@@ -143,7 +135,6 @@ rm_white_multiple <- hijack(rm_default, pattern = "@rm_white_multiple")
 #' endmark (period, question mark, or exclamation point).
 #' 
 #' @rdname rm_white
-#' @family rm_ functions
 #' @include utils.R
 #' @export
 rm_white_punctuation <- hijack(rm_default, pattern = "@rm_white_punctuation")